requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.0.1

opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
scipy==1.12.0
scikit-image==0.22.0
scikit-learn==1.4.0

torch==2.2.0
torchvision==0.17.0
facenet-pytorch==2.5.3

matplotlib==3.8.2
plotly==5.18.0
pandas==2.2.0
